+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Refactored ExtendedInstaller class to be more stable
+- Refactored ExtendedInstaller class to be more reliable
 - Extended resources installer page is now more user friendly
+- Small fixes and code enhancements
 
 ## [0.8.2] - 2020-02-08
 ### Added