+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Refactored ExtendedInstaller class to be more reliable
 - Extended resources installer page is now more user friendly
 - Fixed product price JSON output if a currency field is missing
+- Updated products package installer to add specific module config on install
 - Small fixes and code enhancements
 
 ## [0.8.2] - 2020-02-08